format, lint, move types
