+++ src/features/inbox/types.ts
@@ -11,29 +11,3 @@ export const notificationPayloadSchema = z.object({
 });
 
 export type NotificationPayload = z.infer<typeof notificationPayloadSchema>;
-
-
-// A device group referenced by a notification. All fields are optional so the
-// model can emit whatever identifiers it finds; downstream code skips entries
-// with no usable identifier.
-// TODO: add new fields like UDI after VW-283 gets merged in
-// TODO: add new fields like versionRange after VW-283 gets merged in (used to link individual assets?)
-// vers schema, if we provide that + maybe a skill to use it if necessary, has a way to provide multiple OR versions
-//  https://www.packageurl.org/docs/vers/schemas
-// TODO: if we can find more data, add something like serialRange
-// TODO: What about more unique ID's for specific vendors? e.g, Siemens has material number as a unique device group code
-export const extractedDeviceGroupSchema = z.object({
-  cpe: z.string().nullish(),
-  manufacturer: z.string().nullish(),
-  modelName: z.string().nullish(),
-  version: z.string().nullish(),
-});
-
-// TODO: extend this, which has just device groups for now, with vulnerabilities
-// and remediations
-export const extractSchema = z.object({
-  deviceGroups: z.array(extractedDeviceGroupSchema),
-});
-
-export type ExtractedDeviceGroup = z.infer<typeof extractedDeviceGroupSchema>;
-export type ExtractResult = z.infer<typeof extractSchema>;
