+++ src/lib/s3.ts
@@ -1,5 +1,9 @@
 import "server-only";
-import { GetObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
+import {
+  GetObjectCommand,
+  PutObjectCommand,
+  S3Client,
+} from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 import type { z } from "zod";
 import type { artifactInputSchema } from "@/features/artifacts/types";
